Remove the "Install()" method from the part interface
Move the download of remotesnaps into the SnapUbuntuStoreRepository
