+++ snap/removed/removed.go
@@ -129,11 +129,6 @@ func (r *Removed) DownloadSize() int64 {
 	return -1
 }
 
-// Install from the snappy.Part interface
-func (r *Removed) Install(pb progress.Meter, flags snappy.InstallFlags) (name string, err error) {
-	return "", ErrRemoved
-}
-
 // Config from the snappy.Part interface
 func (r *Removed) Config(configuration []byte) (newConfig string, err error) { return "", ErrRemoved }
