+++ snap/removed/removed_test.go
@@ -87,9 +87,7 @@ func (s *removedSuite) TestNoStore(c *check.C) {
 
 	prog := &progress.NullProgress{}
 	c.Check(part.SetActive(true, prog), check.Equals, ErrRemoved)
-	_, err := part.Install(prog, 0)
-	c.Check(err, check.Equals, ErrRemoved)
-	_, err = part.Config(nil)
+	_, err := part.Config(nil)
 	c.Check(err, check.Equals, ErrRemoved)
 	_, err = part.Frameworks()
 	c.Check(err, check.Equals, ErrRemoved)
@@ -123,9 +121,7 @@ func (s *removedSuite) TestWithStore(c *check.C) {
 
 	prog := &progress.NullProgress{}
 	c.Check(part.SetActive(true, prog), check.Equals, ErrRemoved)
-	_, err := part.Install(prog, 0)
-	c.Check(err, check.Equals, ErrRemoved)
-	_, err = part.Config(nil)
+	_, err := part.Config(nil)
 	c.Check(err, check.Equals, ErrRemoved)
 	_, err = part.Frameworks()
 	c.Check(err, check.Equals, ErrRemoved)
