+++ snappy/parts.go
@@ -101,8 +101,6 @@ type Part interface {
 	InstalledSize() int64
 	DownloadSize() int64
 
-	// Install the snap
-	Install(pb progress.Meter, flags InstallFlags) (name string, err error)
 	// Config takes a yaml configuration and returns the full snap
 	// config with the changes. Note that "configuration" may be empty.
 	Config(configuration []byte) (newConfig string, err error)
